% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_ciw_sample}
\alias{calc_ciw_sample}
\title{Calculates Pre-Posterior Distribution of 95\% CI Widths Using Sampling-based Simulation}
\usage{
calc_ciw_sample(N, parms)
}
\arguments{
\item{N}{A vector of sample sizes}

\item{parms}{Parameters for the distribution containing:
prev: prevalence
dist_type: distribution type
dist_parm1: first parameter of distribution
dist_parm2: second parameter of distribution
cal_int: calibration intercept
cal_slp: calibration slope}
}
\value{
List of length N, of vectors containing 95\% confidence interval width for each of:
cstat: c-statistic
cal_oe: observed to expected ratio
cal_mean: mean calibration
cal_int: calibration intercept
cal_slp: calibration slope
}
\description{
Calculates pre-posterior distribution of 95\% CI widths using sampling-based simulation
}
