% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_se_sp}
\alias{calc_se_sp}
\title{Calculates the Sensitivity and Specificity}
\usage{
calc_se_sp(dist_type, dist_parms, cal_int, cal_slp, threshold, prev)
}
\arguments{
\item{dist_type}{The distribution type, one of c("logitnorm", "beta", "probitnorm").}

\item{dist_parms}{Vector of the two parameters of interest given the distribution.}

\item{cal_int}{The calibration intercept.}

\item{cal_slp}{The calibration slope.}

\item{threshold}{The risk threshold}

\item{prev}{The outcome prevalence, the expectation of the model}
}
\value{
A vector containing sensitivity and specificity
}
\description{
Calculate the sensitivity and specificity of the model at given threshold
}
\examples{
calc_se_sp("beta", c(1,1), 0.9, 0.75, 0.5, 0.5)
}
