% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{infer_correlation}
\alias{infer_correlation}
\title{Calculates Correlation}
\usage{
infer_correlation(dist_type, dist_parms, cal_int, cal_slp, n, n_sim)
}
\arguments{
\item{dist_type}{The distribution type}

\item{dist_parms}{The two parameters of interest for the given distribution type}

\item{cal_int}{The calibration intercept.}

\item{cal_slp}{The calibration slope.}

\item{n}{number of observations for each simulation.}

\item{n_sim}{number of simulations}
}
\value{
correlation among the simulated data
}
\description{
Calculates correlation based on simulated data
}
