% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_cal_instability}
\alias{plot_cal_instability}
\title{Plots Calibration Instability from Simulated Calibration Curves}
\usage{
plot_cal_instability(N, sample, method = "loess", X = (1:99)/100)
}
\arguments{
\item{N}{Number of observations to simulate in each sample}

\item{sample}{Data frame with columns:
dist_type: distribution type
dist_parm1: first distribution parameter (e.g. mean, alpha, shape1)
dist_parm2: second distribution parameter (e.g. sd, beta, shape2)
cal_int: calibration intercept
cal_slp: calibration slope}

\item{method}{One of loess or line, on default is loess}

\item{X}{Vector of predicted probabilities, on default is 0.01 to 0.99}
}
\value{
Plot of simulated calibration curves
}
\description{
Simulates calibration curves based on given method, and uses plot to visualize calibration instability.
}
\examples{
sample <- data.frame(
dist_type = rep("beta", 3),
dist_parm1 = c(1,2,3),
dist_parm2 = c(3,4,5),
cal_int = c(0, 0.05, 0.1),
cal_slp = c(1, 0.9, 0.8))
plot_cal_instability(N=200, sample=sample)
}
