% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{process_evidence_element}
\alias{process_evidence_element}
\title{Creates a clean list including type, parms, moments}
\usage{
process_evidence_element(element)
}
\arguments{
\item{element}{named list with type (e.g "beta", "norm", logitnorm"), and one of valid combination of parameters:
\itemize{
\item "mean" and "var"
\item "m" and "v"
\item "mean" and "sd"
\item "m" and "sd"
\item "alpha" and "beta"
\item "mean" and "cih"
\item "m" and "cih"
}}
}
\value{
an element with type, params and moments.
}
\description{
standardizes element names, and creates a clean list of names
}
\examples{
process_evidence_element(list(type="norm", mean=0, sd=1))
}
