% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayespmtools-package.R
\docType{package}
\name{bayespmtools-package}
\alias{bayespmtools}
\alias{bayespmtools-package}
\title{bayespmtools: Bayesian Approach to Clinical Prediction Model's Development and Validation}
\description{
Bayesian power, sample size, and value-analysis calculator for risk prediction models.
}
\author{
\strong{Maintainer}: Mohsen Sadatsafavi \email{mohsen.sadatsafavi@ubc.ca} (\href{https://orcid.org/0000-0002-0419-7862}{ORCID})

Other contributors:
\itemize{
  \item Anna Luo \email{annaluo@student.ubc.ca} [contributor]
}

}
\keyword{internal}
