% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valprec.R
\name{bpm_valprec}
\alias{bpm_valprec}
\title{Bayesian Precision/VoI Calculator}
\usage{
bpm_valprec(
  N,
  evidence,
  targets = list(fciw.cstat = T, fciw.cal_slp = TRUE, eciw.cstat = TRUE, eciw.cal_oe =
    TRUE, qciw.cal_oe = c(0.9), assurance.nb = TRUE, voi.nb = TRUE),
  n_sim = 1000,
  method = "sample",
  threshold = NULL,
  dist_type = "logitnorm",
  impute_cor = TRUE,
  ex_args = NULL
)
}
\arguments{
\item{N}{Numeric vector of sample sizes to evaluate}

\item{evidence}{A named list with prior evidence parameters}

\item{targets}{list of metrics to compute:
fciw.cstat, fciw.cal_slp: Boolean value for frequentist CI width for cstat and calibration slope. Default is True
eciw.cstat, eciw.cal_oe: Boolean value for expected CI width for cstat, and observed-to-expected ratio. Default it True
qciw.cal_oe: Vector of quantile of CI width for observed-to-expected ratio. Default is 0.9
assurance.nb: Boolean value for Net benefit assurance
voi.nb: Boolean value for Value of Information for net benefit}

\item{n_sim}{number of Monte Carlo simulations. Default is 1000}

\item{method}{method to calculate pre-posterior distribution of 95\% confidence intervals. One of "sample", "2s"; default is "sample"}

\item{threshold}{Threshold used for decision rules and NB calculations. Required if \code{voi.nb} or \code{assurance.nb} are requested.}

\item{dist_type}{Distribution type for prevalence. Default is "logitnorm"}

\item{impute_cor}{Boolean value to induce correlation. Default is True}

\item{ex_args}{List of extra arguments.}
}
\value{
A list containing:
fciw: Frequentist confidence interval widths (if requested)
eciw: Expected posterior CI widths (if requested)
qciw: Quantile posterior CI widths (if requested)
voi: Expected value of information metrics (EVPI, EVSI)
assurance: Net benefit assurance statistics
sample: The full Monte Carlo sample used in computations
evidence: Processed evidence object
}
\description{
Bayesian Precision/VoI calculator for external validation studies of risk prediction models at a given sample size
}
\examples{
evidence <- list(
  prev = list(type = "beta", mean = 0.38, sd = 0.01),  # tight prior for stability
  cstat = list(type="beta", mean = 0.7, sd = 0.05),
  cal_mean = list(mean = 0, sd = 0.3),
  cal_slp = list(mean = 0.8, sd = 0.3)
)

bpm_valprec(
  N = c(1000, 1500),
  evidence = evidence,
  targets = list(fciw.cstat = TRUE),
  impute_cor = FALSE,    # prevent ROC crashes
  n_sim = 100            # faster and safer on CRAN
)

}
