% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valprec.R
\name{bpm_valprec}
\alias{bpm_valprec}
\title{Bayesian Precision/VoI Calculator}
\usage{
bpm_valprec(
  N,
  evidence,
  targets,
  n_sim = NULL,
  method = "sample",
  threshold = NULL,
  dist_type = "logitnorm",
  impute_cor = TRUE,
  ex_args = NULL
)
}
\arguments{
\item{N}{Numeric vector of sample sizes to evaluate}

\item{evidence}{A named list with prior evidence parameters}

\item{targets}{A list of targets to compute:
For expected precision width, the format is eciw.metric=TRUE/FALSE, e.g., eciw.cstat=TRUE will calculate the expected CI width around c-statistic at requested sample sizes.
For assurance, the format is qciw.XXX=q, where q is the desired quantile. For example, qciw.cal_oe=0.9 returns the 90th percentile of CI width around O/E ratio.
Currently implemented metrics are: cstat (c-statistics), cal_oe (observed/expected event ratio: E(Y)/E(pi)), cal_mean (mean calibration error: E(Y)-E(pi)), cal_int (calibration intercept), and cal_slp (calibration slope)
For net benefit, targets are:
oa.nb: if TRUE, will calculate the optimality assurance at requested sample sizes.
voi.nb: if TRUE, will calculate the expected value of sample information (EVSI) at requested sample sizes.}

\item{n_sim}{number of Monte Carlo simulations. Default is 1000}

\item{method}{method to calculate pre-posterior distribution of 95\% confidence intervals. One of "sample", "2s"; default is "sample"}

\item{threshold}{Threshold used for decision rules and NB calculations. Required if \code{voi.nb} or \code{assurance.nb} are requested.}

\item{dist_type}{Distribution type for prevalence. Default is "logitnorm"}

\item{impute_cor}{Boolean value to induce correlation. Default is True}

\item{ex_args}{List of extra arguments.}
}
\value{
A list containing:
results: a matrix including the requested precision / voi metrics for each sample requested sample size.
sample: The full Monte Carlo sample used in computations
evidence: Processed evidence object
targets: same as the corresponding input parameter
ciws: a list containing simulated ciws for each requested metric
}
\description{
Bayesian Precision/VoI calculator for external validation studies of risk prediction models at a given sample size
}
\examples{
evidence <- list(
  prev = list(type = "beta", mean = 0.38, sd = 0.01),  # tight prior for stability
  cstat = list(type="beta", mean = 0.7, sd = 0.05),
  cal_mean = list(mean = 0, sd = 0.3),
  cal_slp = list(mean = 0.8, sd = 0.3)
)

res <- bpm_valprec(
  N = c(1000, 1500),
  evidence = evidence,
  targets = list(eciw.cstat = TRUE, qciw.cal_slp=0.9, voi.nb=0.8),
  threshold=0.2,
  n_sim = 100            # faster and safer on CRAN. Please increase this value for real-world use.
)

print(res$results)

}
