% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valprec.R
\name{bpm_valprec}
\alias{bpm_valprec}
\title{Bayesian Precision / VoI Calculator}
\usage{
bpm_valprec(
  N,
  evidence,
  targets,
  n_sim = NULL,
  method = "sample",
  threshold = NULL,
  dist_type = "logitnorm",
  impute_cor = TRUE,
  ex_args = NULL
)
}
\arguments{
\item{N}{Numeric vector of sample sizes to evaluate.}

\item{evidence}{A named list containing prior evidence components for model performance
parameters (e.g., prevalence, discrimination, calibration).
Alternatively, \code{evidence} may be a data frame of pre-posterior
draws (element \code{$sample}) returned by a previous call to this
function or to \code{bpm_valprec()}, in which case those draws are used
directly.}

\item{targets}{A named list of targets to compute.
\describe{
\item{eciw.metric}{Logical; compute expected CI width.}
\item{qciw.metric}{Numeric scalar in (0,1); CI width quantile.}
\item{oa.nb}{Logical; compute optimality assurance for net benefit.}
\item{voi.nb}{Logical; compute EVSI and EVSI/EVPI.}
}}

\item{n_sim}{#' Number of Monte Carlo simulations used to generate the pre-posterior
distribution. If evidence is a data frame from previous calls to relevant functions,
n_sim will automatically be set to the number of rows of the data frame.}

\item{method}{Method to compute CI widths. One of \code{"sample"} or
\code{"2s"}.}

\item{threshold}{Decision threshold for net benefit calculations.
Required if \code{oa.nb} or \code{voi.nb} are requested.}

\item{dist_type}{Distribution for calibrated risks. Default is
\code{"logitnorm"}.}

\item{impute_cor}{Logical; whether to induce correlation between
parameters.}

\item{ex_args}{Optional list of extra arguments. May include
\code{f_progress}, a custom progress function.}
}
\value{
A list with elements:
\describe{
\item{results}{Matrix of requested metrics by sample size.}
\item{sample}{Monte Carlo sample used for computations.}
\item{evidence}{Processed evidence object.}
\item{targets}{Targets as supplied by the user.}
\item{ciws}{Simulated CI widths for requested metrics.}
}
}
\description{
Bayesian precision and value-of-information calculator for external
validation studies of risk prediction models at fixed sample sizes.
}
\examples{
evidence <- list(
 prev~beta(mean=0.428, sd=0.030),
 cstat~beta(mean=0.761, cih=0.773),
 cal_mean~norm(-0.009, 0.125),  #mean and SD
 cal_slp~norm(0.995, 0.024)     #mean and SD
)

res <- bpm_valprec(
  N = c(1000, 1500),
  evidence = evidence,
  targets = list(eciw.cstat = TRUE, qciw.cal_slp=0.9, voi.nb=0.8),
  threshold=0.2,
  n_sim = 100            # faster and safer on CRAN. Please increase this value for real-world use.
)

print(res$results)

}
