% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valsamp.R
\name{bpm_valsamp}
\alias{bpm_valsamp}
\title{Bayesian Sample Size Calculator}
\usage{
bpm_valsamp(
  evidence,
  targets,
  n_sim = NULL,
  method = "sample",
  threshold = NULL,
  dist_type = "logitnorm",
  impute_cor = TRUE,
  ex_args = NULL
)
}
\arguments{
\item{evidence}{A named list containing prior evidence elements.
Alternatively, \code{evidence} can be the pre-posterior draws (\verb{$sample}) returned by previous
calls to this function or to \code{bpm_valprec()}, in which case the method will directly work
with the pre-posterior draws.}

\item{targets}{A list containing targets for sample size calculations.

Optional elements:
\itemize{
\item For precision-based targets, one can focus either on expected 95\\% confidence
interval widths (prefix \code{eciw.}) or on assurance probabilities around CI width
(prefix \code{qciw.}).
\item For example, \code{eciw.cstat = 0.1} indicates a sample size targeting a width of
0.1 for the 95\\% CI around the c-statistic, whereas
\code{qciw.cal_slp = c(0.90, 0.22)} indicates targeting a 90\\% assurance that the CI width
for the calibration slope will not exceed 0.22.
\item For net benefit, targets include \code{oa.nb}, the Optimality Assurance value
(e.g., \code{oa.nb = 0.9} for 90\\% assurance), and \code{voi.nb}, which is the ratio of
the Expected Value of Sample Information (EVSI) to the Expected Value of Perfect
Information (EVPI) (e.g., \code{voi.nb = 0.8}).
}}

\item{n_sim}{Number of Monte Carlo simulations.}

\item{method}{Method used to calculate the pre-posterior distribution of 95\\% confidence intervals.
One of \code{"sample"} or \code{"2s"}; default is \code{"sample"}.}

\item{threshold}{Threshold used for decision rules and net benefit calculations.
Required if \code{voi.nb} or \code{oa.nb} are requested.}

\item{dist_type}{Distribution for calibrated risks; default is \code{"logitnorm"}.}

\item{impute_cor}{Logical value indicating whether correlation should be induced.
Default is \code{TRUE}.}

\item{ex_args}{A list of extra arguments.}
}
\value{
A list containing:
\itemize{
\item \code{results}: Estimated sample sizes needed to meet the targets.
\item \code{sample}: Data frame of simulated samples.
\item \code{evidence}: Processed evidence object.
\item \code{trace}: Trace output from the stochastic root-finding method.
\item \code{targets}: Same as the corresponding input argument.
}
}
\description{
Bayesian sample size calculation for external validation studies of risk prediction models.
}
\examples{
\donttest{
evidence <- list(
  prev = list(type = "beta", mean = 0.4, sd = 0.02),
  cstat = list(mean = 0.75, sd = 0.03),
  cal_mean = list(mean = 0, sd = 0.1),
  cal_slp = list(mean = 1, sd = 0.1)
)

targets <- list(
  eciw.cstat = 0.1,
  qciw.cstat = c(0.9, 0.1),
  oa.nb = 0.8
)

samp <- bpm_valsamp(
  evidence = evidence,
  targets = targets,
  n_sim = 1000,
  threshold = 0.2
)

print(samp$results)
}

}
