% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valsamp.R
\name{bpm_valsamp}
\alias{bpm_valsamp}
\title{Bayesian Sample Size Calculator}
\usage{
bpm_valsamp(
  evidence,
  targets,
  n_sim,
  method = "sample",
  threshold = NULL,
  dist_type = "logitnorm",
  impute_cor = TRUE,
  ex_args = NULL
)
}
\arguments{
\item{evidence}{A named list with prior evidence elements}

\item{targets}{assurance.nb: Boolean value for Net benefit assurance
voi.nb: Boolean value for Value of Information for net benefit}

\item{n_sim}{Number of simulations}

\item{method}{Method to calculate pre-posterior distribution of 95\% confidence intervals. One of "sample", "2s"; default is "sample"}

\item{threshold}{Threshold used for decision rules and NB calculations. Required if \code{voi.nb} or \code{assurance.nb} are requested.}

\item{dist_type}{Distribution for calibrated risks; default is "logitnorm"}

\item{impute_cor}{Boolean value to induce correlation. Default is True}

\item{ex_args}{List of extra arguments.}
}
\value{
A list containing:
N: Estimated sample sizes needed to meet the targets.
sample: Data frame of simulated samples
evidence: Processed evidence object
trace: Trace output from stochastic root finding method
}
\description{
Bayesian sample size calculation for external validation studies of risk prediction models
}
\examples{
\donttest{
evidence <- list(
  prev = list(type = "beta", mean = 0.4, sd = 0.02),        
  cstat = list(mean = 0.75, sd = 0.03),                    
  cal_mean = list(mean = 0, sd = 0.1),                      
  cal_slp = list(mean = 1, sd = 0.1)                       
)

targets <- list(
  eciw.cstat = 0.05,     
  assurance.nb = 0.8       
)

bpm_valsamp(
  evidence = evidence,
  targets = targets,
  n_sim = 1000,
  threshold = 0.2
)
}
}
