% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valsamp.R
\name{bpm_valsamp}
\alias{bpm_valsamp}
\title{Bayesian sample size calculation for risk prediction models}
\usage{
bpm_valsamp(
  evidence,
  targets,
  n_sim,
  method = "sample",
  threshold = NULL,
  dist_type = "logitnorm",
  impute_cor = TRUE,
  ex_args = NULL
)
}
\arguments{
\item{evidence}{A named list with prior evidence elements}

\item{targets}{assurance.nb: Boolean value for Net benefit assurance
voi.nb: Boolean value for Value of Information for net benefit}

\item{n_sim}{Number of simulations}

\item{method}{method to calculate pre-posterior distribution of 95\% confidence intervals. One of "sample", "2s"; default is "sample"}

\item{threshold}{Threshold used for decision rules and NB calculations. Required if \code{voi.nb} or \code{assurance.nb} are requested.}

\item{dist_type}{Distribution for calibrated risks; default is "logitnorm"}

\item{impute_cor}{Boolean value to induce correlation. Default is True}

\item{ex_args}{List of extra arguments.}
}
\value{
a list containing:
N: Estimated sample sizes needed to meet the targets.
sample: Data frame of simulated samples
evidence: Processed evidence object
trace: trace output from stochastic root finding method TODO
}
\description{
Bayesian sample size calculation for external validation studies of risk prediction models
}
\examples{
evidence <- list(
  prev = list(type = "beta", mean = 0.4, sd = 0.02),        # prevalence
  cstat = list(mean = 0.75, sd = 0.03),                     # discrimination (C-statistic)
  cal_mean = list(mean = 0, sd = 0.1),                      # calibration intercept
  cal_slp = list(mean = 1, sd = 0.1)                        # calibration slope
)

# Specify target criteria for sample size calculation
targets <- list(
  eciw.cstat = 0.05,        # expected confidence interval width for c-statistic
  assurance.nb = 0.8        # net benefit assurance target
)

bpm_valsamp(
  evidence = evidence,
  targets = targets,
  n_sim = 200,
  threshold = 0.25
)
}
