% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_ciw_2s}
\alias{calc_ciw_2s}
\title{Calculates Pre-Posterior Distribution of 95\% CI Widths Using Two-step Method}
\usage{
calc_ciw_2s(N, parms)
}
\arguments{
\item{N}{A vector of sample sizes}

\item{parms}{Parameters for the distribution containing:
cal_int: calibration intercept
cal_slp: calibration slope
prev: prevalence
dist_type: distribution type
cstat: c-statistic
dist_type: one of ("logitnorm", "beta, "probitnorm")
dist_parm1: first parameter of the distribution
dist_parm2: second parameter of the distribution}
}
\value{
List of length N, of vectors containing 95\% confidence interval width for each of:
cstat: c-statistic
cal_oe: observed to expected ratio
cal_mean: mean calibration
cal_int: calibration intercept
cal_slp: calibration slope
}
\description{
Calculates pre-posterior distribution of 95\% CI widths using two-step method.
}
