% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_ciw_mc}
\alias{calc_ciw_mc}
\title{Calculates pre-posterior distribution of 95\% CI widths based on given method}
\usage{
calc_ciw_mc(N, parms_sample, method)
}
\arguments{
\item{N}{A vector of sample sizes}

\item{parms_sample}{matrix of parameters for the distribution each row with appropriate parameters:
cstat: c-statistic
prev: prevalence
dist_type: distribution type
dist_parm1: first parameter of distribution
dist_parm2: seconnd parameter of distribution
cal_int: calibration intercept
cal_slp: calibration slope}

\item{method}{method to calculate 95\% confident interval width, one of sample, 2s}
}
\value{
List of matrices each with dimension (number of rows in parms_sample x length N) containing 95\% confidence interval width for each of:
cstat: c-statistic
cal_oe: observed to expected ratio
cal_mean: mean calibration
cal_int: calibration intercept
cal_slp: calibration slope
}
\description{
Calculates pre-posterior distribution of 95\% CI widths based on given method
}
\examples{
parameters <- data.frame(cstat=c(0.75, 0.8), 
                        prev=c(0.2,0.3), 
                        dist_type=c("beta", "beta"), 
                        dist_parm1=c(0.9, 1),
                        dist_parm2=c(1.5,2),
                        cal_int=c(0, 0.5), 
                        cal_slp=c(1,0.9))
calc_ciw_mc(N=c(100,300,500), parms_sample=parameters, method="2s")
}
