% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_ciw_sample}
\alias{calc_ciw_sample}
\title{Calculates pre-posterior distribution of 95\% CI widths using sampling-based simulation}
\usage{
calc_ciw_sample(N, parms)
}
\arguments{
\item{N}{A vector of sample sizes}

\item{parms}{Parameters for the distribution containing:
prev: prevalence
dist_type: distribution type
dist_parm1: first parameter of distribution
dist_parm2: seconnd parameter of distribution
cal_int: calibration intercept
cal_slp: calibration slope}
}
\value{
List of length N, of vectors containing 95\% confidence interval width for each of:
cstat: c-statistic
cal_oe: observed to expected ratio
cal_mean: mean calibration
cal_int: calibration intercept
cal_slp: calibration slope
}
\description{
Calculates pre-posterior distribution of 95\% CI widths using sampling-based simulation
}
\examples{
parameters<-list(prev = 0.2, dist_type="beta", dist_parm1=2, dist_parm2=4, cal_int = 0, cal_slp = 1)
calc_ciw_sample(N=c(100,300,500), parms=parameters)
}
