% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_cstat}
\alias{calc_cstat}
\title{Calculate the c-statistic of the model}
\usage{
calc_cstat(type, parms, m = NULL)
}
\arguments{
\item{type}{A character string; one of c("beta", "logitnorm", "probitnorm") indicating the model type.}

\item{parms}{A numeric vector containing parameters relevant to the model.}

\item{m}{#TODO!}
}
\value{
The c-statistic.
}
\description{
Calculates the c-statistic given the model type and parameters.
}
\examples{
calc_cstat("norm", c(0,2))
}
