% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_riley_vars}
\alias{calc_riley_vars}
\title{Calculates Approximate Variances and Covariance for Performance Metrics}
\usage{
calc_riley_vars(N, parms)
}
\arguments{
\item{N}{sample size of the validation dataset}

\item{parms}{list containing model and distribution parameters:
prev: expected prevalence
cstat: c-statistic of the model
dist_type: one of ("logitnorm", "beta, "probitnorm")
dist_parm1: first parameter of the distribution
dist_parm2: second parameter of the distribution
cal_int: calibration intercept
cal_slp: calibration slope}
}
\value{
list of approximate variances and covariance of the performance metrics.
}
\description{
Calculates approximate variances performance metrics and covariance of calibration intercept and slope using the Riley framework
}
\examples{
parameters <- list(
  prev=0.5, 
  cstat=0.75, 
  dist_type="beta", 
  dist_parm1=1,
  dist_parm2=1, 
  cal_int=0, 
  cal_slp=1)
   
calc_riley_vars(N=100, parms=parameters)
}
