% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_se_sp}
\alias{calc_se_sp}
\title{Calculate the sensitivity and specificity of the model at given threshold}
\usage{
calc_se_sp(dist_type, dist_parms, cal_int, cal_slp, threshold, prev)
}
\arguments{
\item{dist_type}{One of c("logitnorm", "beta", "probitnorm")}

\item{dist_parms}{Vector of the two parameters of interest given the distribution}

\item{cal_int}{calibration intercept}

\item{cal_slp}{calibration slope}

\item{threshold}{risk threshold}

\item{prev}{outcome prevalence, the expectation of the model}
}
\value{
A vector containing sensitivity and specificity
}
\description{
Calculate the sensitivity and specificity of the model at given threshold
}
\examples{
calc_se_sp("beta", c(1,1), 0.9, 0.75, 0.5, 0.5)
}
