% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{find_n_mean}
\alias{find_n_mean}
\title{Calculates Sample Size Given Target Mean CI}
\usage{
find_n_mean(target, N, ciws, decreasing = T, convex = T)
}
\arguments{
\item{target}{The target mean confidence interval width}

\item{N}{Sample sizes corresponding to each row of ciws,=}

\item{ciws}{Matrix of confidence intervals widths, each row corresponding to N}

\item{decreasing}{Logical. Constraining function to decreasing}

\item{convex}{Logical. Constraining function to convex}
}
\value{
Integer. Estimated sample size needed to achieve the target
}
\description{
Calculates sample size N, so that the mean confidence interval is equal to given target, assumes function is decreasing and convex
}
