% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{find_n_quantile}
\alias{find_n_quantile}
\title{Calculates Sample Size Given Target Quantile}
\usage{
find_n_quantile(target, N, q, ciws)
}
\arguments{
\item{target}{The desired quantile target value}

\item{N}{Sample sizes corresponding to each row of ciws}

\item{q}{Desired quantile level, between 0 and 1.}

\item{ciws}{A matrix of confidence intervals widths, each row corresponding to N}
}
\value{
Estimated sample size needed to achieve the target
}
\description{
Find sample size N, so that the specified quantile is equal to given target
}
