% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{infer_cal_int_from_mean}
\alias{infer_cal_int_from_mean}
\title{Infer Calibration Intercept from Mean Calibration}
\usage{
infer_cal_int_from_mean(dist_type, dist_parms, cal_mean, cal_slp, prev = NULL)
}
\arguments{
\item{dist_type}{The distribution type, one of c("logitnorm", "probitnorm", "beta").}

\item{dist_parms}{The two parameters that index the type.}

\item{cal_mean}{The mean calibration.}

\item{cal_slp}{The calibration slope.}

\item{prev}{Outcome prevalence. Optional; if not provided, estimate is as the expected value of the distribution of calibrated risks.}
}
\value{
The estimated calibration intercept
}
\description{
Infer calibration intercept from mean calibration given a fixed calibration slope and a given distribution for calibrated risks
}
\examples{
infer_cal_int_from_mean("beta", c(1,1), 1 ,1.1, 0.25)
}
