% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{infer_correlation}
\alias{infer_correlation}
\title{Calculates correlation based on simulated data}
\usage{
infer_correlation(dist_type, dist_parms, cal_int, cal_slp, n, n_sim)
}
\arguments{
\item{dist_type}{The distribution type}

\item{dist_parms}{The two parameters of interest for the given distribution type}

\item{cal_int}{Calibration intercept}

\item{cal_slp}{Calibration slope}

\item{n}{number of observations for each simulation}

\item{n_sim}{number of simulations}
}
\value{
correlation among the simulated data
}
\description{
calculates correlation based on simulated data
}
\examples{
infer_correlation("beta", c(1,1), 1, 0.5, 10, 10)
}
