% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{inv_mean_quantile}
\alias{inv_mean_quantile}
\title{Calculate the model parameters given the distribution type, mean, quantile, and percentile.}
\usage{
inv_mean_quantile(type, m, q, p)
}
\arguments{
\item{type}{Type of model, one of ("logitnorm", "beta", "probitnorm", "norm")}

\item{m}{mean}

\item{q}{quantile}

\item{p}{percentile}
}
\value{
The two parameters of the given type.
}
\description{
Calculate the model parameters given the distribution type, mean, quantile, and percentile.
}
\examples{
inv_mean_quantile("beta", 0.5, 0.25, 0.25)
}
