% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{inv_mean_quantile}
\alias{inv_mean_quantile}
\title{Calculates the Model Parameters Given Quantile}
\usage{
inv_mean_quantile(type, m, q, p)
}
\arguments{
\item{type}{The distribution type, one of c("norm", "beta", "logitnorm", "probitnorm").}

\item{m}{Mean of the of distribution.}

\item{q}{The quantile value.}

\item{p}{The percentile at which the quantile occurs.}
}
\value{
The model parameters of the given type.
}
\description{
Calculate the model parameters given the distribution type, mean, quantile, and percentile.
}
