% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{inv_moments}
\alias{inv_moments}
\title{Calculate the model parameters given the first two moments.}
\usage{
inv_moments(type, moments)
}
\arguments{
\item{type}{A string of the type of model; one of c("norm", "logitnorm", "beta")}

\item{moments}{A numeric value denoting the first two moments of the model}
}
\value{
#'Returns mean and sd for norm
Returns mu and sigma for logitnorm
Returns shape1 (alpha) and shape2 (beta) for beta
}
\description{
Calculates the model parameters of interest given the first two moments.
}
\examples{
inv_moments("norm", c(0,1))
}
