% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{moments}
\alias{moments}
\title{Calculate mean and variance of a model}
\usage{
moments(type, parms)
}
\arguments{
\item{type}{A string; one of c("norm", "beta", "logitnorm", "probitnorm") indicating the model type.}

\item{parms}{A numeric vector containing parameters relevant to the model.}
}
\value{
A numeric vector representing the mean and variance.
}
\description{
Calculates the first two moments (mean and variance) of the given model type and parameters.
}
\examples{
moments("norm", c(0, 1))
moments("beta", c(1, 1))
}
