% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence.R
\name{process_evidence}
\alias{process_evidence}
\title{Transforms Evidence Into Standardized Format}
\usage{
process_evidence(evidence)
}
\arguments{
\item{evidence}{named list of evidence elements including:
prev: prevalence
cstat: c-statistic
cal_slp: calibration slope and,
one of cal_mean (mean calibration), cal_oe (observed to expected ratio), or cal_int (calibration intercept)}
}
\value{
Modified evidence object that has been standardized and restructured
}
\description{
Verifies evidence object has correct members, and standardizes it
}
\examples{
evidence <- list(
prev=list(type="beta", mean=0.38, sd=0.2),
cstat=list(mean=0.7, sd=0.05),
cal_int=list(mean=0.2, sd=0.2),
cal_slp=list(mean=0.8, sd=0.3))
process_evidence(evidence=evidence)
}
