% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rbnorm}
\alias{rbnorm}
\title{Generates Samples From Normal Distribution}
\usage{
rbnorm(n, mu1, mu2, var1, var2, cov)
}
\arguments{
\item{n}{Number of samples to be generated}

\item{mu1}{Mean of first variable}

\item{mu2}{Mean of second variable}

\item{var1}{Variance of first variable}

\item{var2}{Variance of second variable}

\item{cov}{Covariance between the two variables}
}
\value{
Matrix of nx2 where
column 1 contains samples for the first variable, and
column 2 contains samples for the second variable conditioned on the first
}
\description{
generates samples from a normal distribution using marginal means, variances, and covariance
}
\examples{
rbnorm(n = 1000, mu1 = 0, mu2 = 0, var1 = 1, var2 = 1, cov = 0.5)
}
