% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rbnorm}
\alias{rbnorm}
\title{Generates samples from a normal distribution using marginal means, variances, and covariance}
\usage{
rbnorm(n, mu1, mu2, var1, var2, cov)
}
\arguments{
\item{n}{number of samples to be generated}

\item{mu1}{mean of first variable}

\item{mu2}{mean of second variable}

\item{var1}{variance of first variable}

\item{var2}{variance of second variable}

\item{cov}{covariance between the two variables}
}
\value{
returns matrix of nx2 where
column 1 contains samples for the first variable, and
column 2 contains samples for the second variable conditioned on the first
}
\description{
generates samples from a normal distribution using marginal means, variances, and covariance
}
\examples{
rbnorm(n = 1000, mu1 = 0, mu2 = 0, var1 = 1, var2 = 1, cov = 0.5)
}
