% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{riley_samp}
\alias{riley_samp}
\title{Calculates Sample Size that Achieves Target CI Widths}
\usage{
riley_samp(target_ciws, parms)
}
\arguments{
\item{target_ciws}{Named list containing target confidence interval width for at least one of:
prev: prevalence
cstat: c-statistic
cal_mean: mean calibration
cal_oe: observed to expected outcome ratio
cal_int: calibration intercept
cal_slp: calibration slope}

\item{parms}{List containing model parameters and distribution:
prev: expected prevalence
cstat: c-statistic of the model
dist_type: one of ("logitnorm", "beta, "probitnorm")
dist_parm1: first parameter of the distribution
dist_parm2: second parameter of the distribution
cal_int: calibration intercept
cal_slp: calibration slope}
}
\value{
A named list of estimated sample sizes that achieve target confidence interval widths:
fciw.prev, fciw.cstat, fciw.cal_mean, fciw.cal_oe, fciw.cal_int, fciw.cal_slp
}
\description{
Calculates sample size that achieves target confidence interval widths using Riley's framework
}
