% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{riley_samp}
\alias{riley_samp}
\title{calculates sample size that achieves target confidence interval widths using Riley's framework}
\usage{
riley_samp(target_ciws, parms)
}
\arguments{
\item{target_ciws}{named list containing target confidence interval width for at least one of:
prev: prevalence
cstat: c-statistic
cal_mean: mean calibration
cal_oe: observed to expected outcome ratio
cal_int: calibration intercept
cal_slp: calibration slope}

\item{parms}{list containing model parameters and distribution:
prev: expected prevalence
cstat: c-statistic of the model
dist_type: one of ("logitnorm", "beta, "probitnorm")
dist_parm1: first parameter of the distribution
dist_parm2: second parameter of the distribution
cal_int: calibration intercept
cal_slp: calibration slope}
}
\value{
a named list of estimated sample sizes that achieve target confidence interval widths:
fciw.prev, fciw.cstat, fciw.cal_mean, fciw.cal_oe, fciw.cal_int, fciw.cal_slp
}
\description{
calculates sample size that achieves target confidence interval widths using Riley's framework
}
\examples{
parameters <- list(
  prev=0.5, 
  cstat=0.75, 
  dist_type="beta", 
  dist_parm1=1, 
  dist_parm2=1, 
  cal_int=0, 
  cal_slp=1)
  
riley_samp(target_ciws=list(prev=0.05, cstat=0.05) ,parms=parameters)
}
